Move component to correct directory
refs #47
