+++ app/listing/[id]/_component/main-info-col.tsx
