+++ app/listing/[id]/_component/main-info.tsx
