+++ app/listing/[id]/_component/photos.tsx
