+++ stories/_app/ListingPageDescription.stories.tsx
@@ -1,4 +1,4 @@
-import ListingPageDescription from "@/app/listing/_component/description";
+import ListingPageDescription from "@/app/listing/[id]/_component/description";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof ListingPageDescription> = {
